pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yolov8-safety-detection"
version = "0.1.0"
description = "YOLOv8-powered worker safety equipment detection system"
readme = "docs/README.md"
requires-python = ">=3.10"
authors = [
    {name = "A-Kuo", email = "akuo@example.com"}
]
keywords = ["yolo", "safety", "detection", "edge-ai", "qualcomm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "ultralytics>=8.0.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.17.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "isort>=5.12.0",
]

qnn = [
    "onnxruntime-qnn>=1.17.0",
]

[project.urls]
Repository = "https://github.com/a-kuo/yolov8-powered-safety-equipment-detection-system"
Issues = "https://github.com/a-kuo/yolov8-powered-safety-equipment-detection-system/issues"

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --strict-markers"
markers = [
    "slow: marks tests as slow",
    "gpu: marks tests that require GPU",
]
